les2_autogen/homework/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.10.0
pydantic-settings>=2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
autogen-agentchat==0.7.4
autogen-ext[openai]==0.7.4
autogen-core==0.7.4
openai>=1.93
httpx==0.25.2
websockets==12.0
sse-starlette==1.8.2
